Rework lib/(proxy-)api-depot
