+++ lib/api-depot.js
@@ -1,6 +1,9 @@
+const API_DEPOT_URL = process.env.NEXT_PUBLIC_API_DEPOT_URL || 'https://plateforme.adresse.data.gouv.fr/api-depot'
+
 async function _fetch(url, method, body) {
   const options = {
-    method
+    method,
+    mode: 'cors'
   }
 
   if (method === 'POST') {
@@ -10,11 +13,13 @@ async function _fetch(url, method, body) {
     }
   }
 
-  const response = await fetch(url, options)
+  const response = await fetch(API_DEPOT_URL + url, options)
   const contentType = response.headers.get('content-type')
 
   if (!response.ok) {
-    throw new Error(response.message)
+    const error = new Error(response.message)
+    error.code = 404
+    throw error
   }
 
   if (response.ok && contentType && contentType.includes('application/json')) {
@@ -24,53 +29,23 @@ async function _fetch(url, method, body) {
   throw new Error('Une erreur est survenue')
 }
 
-async function uploadFile(revisionId, file) {
-  try {
-    return fetch(`/revisions/${revisionId}/files/bal`, {
-      method: 'PUT',
-      headers: {'Content-Type': 'text/csv'},
-      body: file
-    })
-  } catch (error) {
-    throw new Error(`Impossible de téléverser le fichier : ${error}`)
-  }
-}
-
-async function computeRevision(revisionId) {
-  return _fetch(`/revisions/${revisionId}/compute`, 'POST')
-}
-
-export async function publishRevision(revisionId, body) {
-  return _fetch(`/revisions/${revisionId}/publish`, 'POST', body)
-}
-
-export async function sendAuthenticationCode(habilitationId) {
-  return _fetch(`/habilitations/${habilitationId}/authentication/email/send-pin-code`, 'POST')
-}
-
-export async function submitAuthentificationCode(habilitationId, code) {
-  return _fetch(`/habilitations/${habilitationId}/authentication/email/validate-pin-code`, 'POST', {code})
-}
-
 export async function getRevision(revisionId) {
   return _fetch(`/revisions/${revisionId}`)
 }
 
+export async function getCurrentRevision(codeCommune) {
+  try {
+    const currentRevision = _fetch(`/communes/${codeCommune}/current-revision`)
+    return currentRevision
+  } catch (error) {
+    if (error.code === 404) {
+      return
+    }
+
+    throw error
+  }
+}
+
 export async function getRevisions(codeCommune) {
   return _fetch(`/communes/${codeCommune}/revisions`)
-}
-
-export async function createRevision(codeCommune, context, file) {
-  const revision = await _fetch(`/communes/${codeCommune}/revisions`, 'POST', {context})
-  await uploadFile(revision._id, file)
-
-  return computeRevision(revision._id)
-}
-
-export async function getHabilitation(habilitationId) {
-  return _fetch(`/habilitations/${habilitationId}`)
-}
-
-export async function createHabilitation(codeCommune) {
-  return _fetch(`/communes/${codeCommune}/habilitations`, 'POST')
 }
